requirements.txt
torch
opencv-python
mlflow
ultralytics
streamlit
supervision
mlflow
matplotlib-inline
ipykernel
typer
python-dotenv